Removed football.json from .gitignore
Since football.json is already in the repo, putting it .gitignore doesn't help. I'll just have to be careful not to commit a new football.json if I keep editing code, because it causes the automatic "git pull" on the Pi to fail because it has its own football.json it doesn't want to overwrite.